requirements.txt
fastapi
uvicorn
pydantic
supabase-py
python-dotenv
pynacl
solana
discord.py
requests
solders
